Move `distributeSurplusSpec`.
